# api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Riftbound: Echoes of the Aether API",
    "description": "Server-Authoritative Anti-Cheat & Live-Ops API for Riftbound: Echoes of the Aether",
    "version": "1.0.0-PROD"
  },
  "paths": {
...
    "/api/v1/rift/start": {
      "post": {
        "tags": [
          "Rift Simulation & Anti-Cheat"
        ],
        "summary": "Start Rift Run",
        "operationId": "start_rift_run_api_v1_rift_start_post",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RiftStartRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RiftStartResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/rift/floor/complete": {
      "post": {
        "tags": [
          "Rift Simulation & Anti-Cheat"
        ],
        "summary": "Complete Floor",
        "operationId": "complete_floor_api_v1_rift_floor_complete_post",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FloorCompleteRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FloorCompleteResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/rift/extract": {
      "post": {
        "tags": [
          "Rift Simulation & Anti-Cheat"
        ],
        "summary": "Extract Rift Run",
        "operationId": "extract_rift_run_api_v1_rift_extract_post",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RiftExtractRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RiftExtractResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "CombatMetricsPayload": {
        "properties": {
          "enemies_slain": {
            "type": "integer",
            "title": "Enemies Slain"
          },
          "damage_dealt": {
            "type": "integer",
            "title": "Damage Dealt"
          },
          "damage_taken": {
            "type": "integer",
            "title": "Damage Taken"
          },
          "resonances_triggered": {
            "type": "integer",
            "title": "Resonances Triggered"
          },
          "clear_time_ms": {
            "type": "integer",
            "title": "Clear Time Ms"
          },
          "client_signature_hash": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Client Signature Hash"
          }
        },
        "type": "object",
        "required": [
          "enemies_slain",
          "damage_dealt",
          "damage_taken",
          "resonances_triggered",
          "clear_time_ms"
        ],
        "title": "CombatMetricsPayload"
      },
      "FloorCompleteRequest": {
        "properties": {
          "session_id": {
            "type": "string",
            "title": "Session Id"
          },
          "floor_number": {
            "type": "integer",
            "title": "Floor Number"
          },
          "combat_metrics": {
            "$ref": "#/components/schemas/CombatMetricsPayload"
          },
          "selected_augment_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Selected Augment Id"
          }
        },
        "type": "object",
        "required": [
          "session_id",
          "floor_number",
          "combat_metrics"
        ],
        "title": "FloorCompleteRequest"
      },
      "FloorCompleteResponse": {
        "properties": {
          "validated": {
            "type": "boolean",
            "title": "Validated"
          },
          "status_message": {
            "type": "string",
            "title": "Status Message"
          },
          "next_floor": {
            "type": "integer",
            "title": "Next Floor"
          },
          "score_delta": {
            "type": "integer",
            "title": "Score Delta"
          },
          "total_score": {
            "type": "integer",
            "title": "Total Score"
          },
          "pending_aether": {
            "type": "integer",
            "title": "Pending Aether"
          },
          "pending_void_fragments": {
            "type": "integer",
            "title": "Pending Void Fragments"
          },
          "risk_options": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Risk Options"
          }
        },
        "type": "object",
        "required": [
          "validated",
          "status_message",
          "next_floor",
          "score_delta",
          "total_score",
          "pending_aether",
          "pending_void_fragments",
          "risk_options"
        ],
        "title": "FloorCompleteResponse"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "RiftExtractRequest": {
        "properties": {
          "session_id": {
            "type": "string",
            "title": "Session Id"
          }
        },
        "type": "object",
        "required": [
          "session_id"
        ],
        "title": "RiftExtractRequest"
      },
      "RiftExtractResponse": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "final_score": {
            "type": "integer",
            "title": "Final Score"
          },
          "floors_cleared": {
            "type": "integer",
            "title": "Floors Cleared"
          },
          "aether_awarded": {
            "type": "integer",
            "title": "Aether Awarded"
          },
          "void_fragments_awarded": {
            "type": "integer",
            "title": "Void Fragments Awarded"
          },
          "mastery_xp_gained": {
            "type": "integer",
            "title": "Mastery Xp Gained"
          },
          "new_account_level": {
            "type": "integer",
            "title": "New Account Level"
          }
        },
        "type": "object",
        "required": [
          "success",
          "status",
          "final_score",
          "floors_cleared",
          "aether_awarded",
          "void_fragments_awarded",
          "mastery_xp_gained",
          "new_account_level"
        ],
        "title": "RiftExtractResponse"
      },
      "RiftStartRequest": {
        "properties": {
          "hero_id": {
            "type": "string",
            "title": "Hero Id",
            "default": "hero_kaelen"
          },
          "rift_tier": {
            "type": "integer",
            "title": "Rift Tier",
            "default": 1
          },
          "selected_weapon_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Selected Weapon Id"
          }
        },
        "type": "object",
        "title": "RiftStartRequest"
      },
      "RiftStartResponse": {
        "properties": {
          "session_id": {
            "type": "string",
            "title": "Session Id"
          },
          "seed": {
            "type": "integer",
            "title": "Seed"
          },
          "hero_id": {
            "type": "string",
            "title": "Hero Id"
          },
          "rift_tier": {
            "type": "integer",
            "title": "Rift Tier"
          },
          "initial_floor": {
            "type": "integer",
            "title": "Initial Floor"
          },
          "biome": {
            "type": "string",
            "title": "Biome"
          },
          "starting_health": {
            "type": "integer",
            "title": "Starting Health"
          },
          "floor_hazards": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Floor Hazards"
          }
        },
        "type": "object",
        "required": [
          "session_id",
          "seed",
          "hero_id",
          "rift_tier",
          "initial_floor",
          "biome",
          "starting_health",
          "floor_hazards"
        ],
        "title": "RiftStartResponse"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}